Enhance js-debug debugging configuration (#636)
* Add jsDebugConfigAdapter
