+++ src/debugger/requestArgs.ts
@@ -1,0 +1,57 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for details.
+
+import { ISourceMapPathOverrides } from "./jsDebugConfigAdapter";
+import { DebugProtocol } from "vscode-debugprotocol";
+
+export interface ICordovaAttachRequestArgs extends DebugProtocol.AttachRequestArguments {
+    cwd: string; /* Automatically set by VS Code to the currently opened folder */
+    port: number;
+    url?: string;
+    address?: string;
+    trace?: string;
+    timeout: number;
+    platform: string;
+    target?: string;
+    envFile?: string;
+    env?: any;
+    allEnv?: any;
+    skipFiles?: [];
+    sourceMaps?: boolean;
+    sourceMapPathOverrides?: ISourceMapPathOverrides;
+    webkitRangeMin?: number;
+    webkitRangeMax?: number;
+    attachAttempts?: number;
+    attachDelay?: number;
+    attachTimeout?: number;
+    simulatorInExternalBrowser?: boolean;
+
+    // Ionic livereload properties
+    ionicLiveReload?: boolean;
+    devServerPort?: number;
+    devServerAddress?: string;
+
+    // Cordova-simulate properties
+    simulatePort?: number;
+    livereload?: boolean;
+}
+
+export interface ICordovaLaunchRequestArgs extends DebugProtocol.LaunchRequestArguments, ICordovaAttachRequestArgs {
+    iosDebugProxyPort?: number;
+    appStepLaunchTimeout?: number;
+
+    // Ionic livereload properties
+    devServerTimeout?: number;
+
+    // Chrome debug properties
+    userDataDir?: string;
+    runtimeExecutable?: string;
+    runtimeArgs?: string[];
+
+    // Cordova-simulate properties
+    forceprepare?: boolean;
+    simulateTempDir?: string;
+    corsproxy?: boolean;
+    runArguments?: string[];
+    cordovaExecutable?: string;
+}
